# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Finwise API",
    "version": "0.1.0",
    "description": "First vertical-slice contract for identity, workspace and immutable ledger flows."
  },
  "servers": [{ "url": "/v1" }],
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/health": {
      "get": {
        "operationId": "getHealth",
        "security": [],
        "responses": {
          "200": {
            "description": "Service is ready",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HealthResponse" }
              }
            }
          }
        }
      }
    },
    "/session/bootstrap": {
      "get": {
        "operationId": "bootstrapSession",
        "responses": {
          "200": {
            "description": "Identity and accessible workspaces",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BootstrapResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/workspaces": {
      "post": {
        "operationId": "createWorkspace",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workspace created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Workspace" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/workspaces/{workspaceId}/overview": {
      "get": {
        "operationId": "getWorkspaceOverview",
        "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }],
        "responses": {
          "200": {
            "description": "Permission-filtered overview",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OverviewResponse" }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/workspaces/{workspaceId}": {
      "get": {
        "operationId": "getWorkspace",
        "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }],
        "responses": {
          "200": {
            "description": "Workspace",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Workspace" }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/workspaces/{workspaceId}/accounts": {
      "get": {
        "operationId": "listAccounts",
        "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }],
        "responses": {
          "200": {
            "description": "Visible accounts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Account" }
                }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "operationId": "createAccount",
        "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateAccountRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Account" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/workspaces/{workspaceId}/accounts/{accountId}": {
      "get": {
        "operationId": "getAccount",
        "parameters": [
          { "$ref": "#/components/parameters/WorkspaceId" },
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Visible account",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Account" }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/workspaces/{workspaceId}/transactions": {
      "get": {
        "operationId": "listTransactions",
        "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }],
        "responses": {
          "200": {
            "description": "Visible immutable journals",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Transaction" }
                }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "operationId": "postTransaction",
        "parameters": [
          { "$ref": "#/components/parameters/WorkspaceId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostTransactionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Journal posted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Transaction" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/workspaces/{workspaceId}/transactions/{transactionId}": {
      "get": {
        "operationId": "getTransaction",
        "parameters": [
          { "$ref": "#/components/parameters/WorkspaceId" },
          {
            "name": "transactionId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Visible immutable journal",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Transaction" }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "WorkspaceId": {
        "name": "workspaceId",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": { "type": "string", "minLength": 1, "maxLength": 255 }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
      "Forbidden": {
        "description": "Permission denied",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
      "ValidationError": {
        "description": "Validation failed",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
      "Conflict": {
        "description": "Command conflict",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      }
    },
    "schemas": {
      "MoneyDto": {
        "type": "object",
        "required": ["currency", "minorUnits"],
        "properties": {
          "currency": { "type": "string", "const": "VND" },
          "minorUnits": { "type": "string", "pattern": "^-?(0|[1-9][0-9]*)$" }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" },
          "details": {
            "type": "object",
            "additionalProperties": { "type": "string" }
          },
          "requestId": { "type": "string" }
        }
      },
      "HealthResponse": {
        "type": "object",
        "required": ["status", "service"],
        "properties": {
          "status": { "type": "string", "const": "ok" },
          "service": { "type": "string", "const": "finwise-api" }
        }
      },
      "User": {
        "type": "object",
        "required": ["id", "displayName"],
        "properties": {
          "id": { "type": "string" },
          "displayName": { "type": "string" },
          "email": { "type": "string", "format": "email" }
        }
      },
      "Workspace": {
        "type": "object",
        "required": ["id", "name", "intent", "currency", "status"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "kind": { "type": "string" },
          "intent": { "type": "string", "enum": ["PERSONAL", "SHARED"] },
          "currency": { "type": "string", "const": "VND" },
          "status": { "type": "string" }
        }
      },
      "BootstrapResponse": {
        "type": "object",
        "required": ["user", "workspaces", "suggestedWorkspaceId"],
        "properties": {
          "user": { "$ref": "#/components/schemas/User" },
          "workspaces": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Workspace" }
          },
          "suggestedWorkspaceId": { "type": "string" }
        }
      },
      "CreateWorkspaceRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "kind": {
            "type": "string",
            "enum": ["personal", "family", "class_fund", "other"]
          }
        }
      },
      "CreateAccountRequest": {
        "type": "object",
        "required": ["name", "kind"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "kind": {
            "type": "string",
            "enum": [
              "cash",
              "bank",
              "savings",
              "investment_cash",
              "loan_receivable",
              "liability"
            ]
          },
          "visibilityMode": { "type": "string" }
        }
      },
      "Account": {
        "type": "object",
        "required": [
          "id",
          "workspaceId",
          "name",
          "kind",
          "currency",
          "balanceMinorUnits",
          "visibilityMode"
        ],
        "properties": {
          "id": { "type": "string" },
          "workspaceId": { "type": "string" },
          "name": { "type": "string" },
          "kind": { "type": "string" },
          "currency": { "type": "string", "const": "VND" },
          "balanceMinorUnits": { "type": "string" },
          "visibilityMode": { "type": "string" }
        }
      },
      "PostTransactionRequest": {
        "type": "object",
        "oneOf": [
          { "required": ["type", "amountMinorUnits", "accountId"] },
          {
            "required": [
              "type",
              "amountMinorUnits",
              "sourceAccountId",
              "destinationAccountId"
            ]
          }
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": ["income", "expense", "transfer"]
          },
          "amountMinorUnits": { "type": "string" },
          "accountId": { "type": "string" },
          "sourceAccountId": { "type": "string" },
          "destinationAccountId": { "type": "string" },
          "effectiveDate": { "type": "string", "format": "date" },
          "description": { "type": "string" }
        }
      },
      "TransactionEntry": {
        "type": "object",
        "required": ["id", "accountId", "amountMinorUnits", "direction"],
        "properties": {
          "id": { "type": "string" },
          "accountId": { "type": "string" },
          "amountMinorUnits": { "type": "string" },
          "direction": { "type": "string", "enum": ["increase", "decrease"] }
        }
      },
      "Transaction": {
        "type": "object",
        "required": [
          "id",
          "workspaceId",
          "kind",
          "status",
          "amount",
          "effectiveDate",
          "recordedAt",
          "entries"
        ],
        "properties": {
          "id": { "type": "string" },
          "workspaceId": { "type": "string" },
          "kind": { "type": "string" },
          "status": { "type": "string" },
          "amount": { "$ref": "#/components/schemas/MoneyDto" },
          "effectiveDate": { "type": "string", "format": "date" },
          "recordedAt": { "type": "string", "format": "date-time" },
          "description": { "type": "string" },
          "reversalOfId": { "type": "string" },
          "entries": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TransactionEntry" }
          }
        }
      },
      "OverviewResponse": {
        "type": "object",
        "required": [
          "workspaceId",
          "period",
          "accounts",
          "recentTransactions",
          "budgets",
          "totals",
          "hasPartialAccess"
        ],
        "properties": {
          "workspaceId": { "type": "string" },
          "period": { "type": "string" },
          "accounts": { "type": "array", "items": { "type": "object" } },
          "recentTransactions": {
            "type": "array",
            "items": { "type": "object" }
          },
          "budgets": { "type": "array", "items": { "type": "object" } },
          "totals": { "type": "object" },
          "hasPartialAccess": { "type": "boolean" }
        }
      }
    }
  }
}